+++ lib/deploy.command.js
@@ -1,8 +1,8 @@
 const path = require('path')
 const fs = require('fs')
 
-const { readFile, fileExist, mkdirp, symlink } = require('./fs.utils')
-const { log } = require('./log.utils')
+const { readFile, fileExist, mkdirp, symlink } = require('./shared/fs.utils')
+const { log } = require('./shared/log.utils')
 
 module.exports = exports = configFilename => {
   return (modules, options) => {
@@ -62,7 +62,7 @@ function createLinks(files) {
         () => successFulLink.push(file.to),
         error => {
           if (error.code !== 'EEXIST') {
-            throw error 
+            throw error
           }
         }
       )
