+++ lib/init.command.js
@@ -2,9 +2,9 @@ const fs = require('fs')
 const path = require('path')
 const inquirer = require('inquirer')
 
-const { fileExist, writeFile } = require('./fs.utils')
-const { log } = require('./log.utils')
-const { gitClone } = require('./git.utils')
+const { fileExist, writeFile } = require('./shared/fs.utils')
+const { log } = require('./shared/log.utils')
+const { gitClone } = require('./shared/git.utils')
 
 const URL_REGEX = /^((?:(https?):\/\/)?((?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9])\.(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9])\.)(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9])\.)(?:(?:25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[0-9][0-9]|[0-9]))|(?:(?:(?:\w+\.){1,2}[\w]{2,3})))(?::(\d+))?((?:\/[\w]+)*)(?:\/|(\/[\w]+\.[\w]{3,4})|(\?(?:([\w]+=[\w]+)&)*([\w]+=[\w]+))?|\?(?:(wsdl|wadl))))$/
 const PATH_REGEX = /^(?:(?:~|\.{1,2})?\/)+(?:[a-zA-Z\.\-\_]+\/?)*$/
@@ -40,7 +40,7 @@ module.exports = exports = configFilename => {
         if (!overwrite_file) {
           throw new Error('user_stop_init')
         }
-        
+
         folder_path = folder_path.replace('~', process.env.HOME)
         folder_path = path.resolve(folder_path)
 
@@ -53,17 +53,17 @@ module.exports = exports = configFilename => {
           log({ type: 'info', message: 'Folder does not exist. It will be created.' })
           fs.mkdirSync(folderPath)
         }
-        
+
         const isDirectory = fs.statSync(folderPath).isDirectory()
         if (!isDirectory) {
           throw new Error('folder_is_not_directory')
         }
-        
+
         const folderIsEmpty = fs.readdirSync(folderPath).length < 1
         if (!folderIsEmpty) {
           throw new Error('folder_is_not_empty')
         }
-        
+
         return data
       })
       .then(data => {
@@ -71,7 +71,7 @@ module.exports = exports = configFilename => {
           log({ type: 'warn', message: 'Impossible to cloned git repository. Need repository URL.' })
           return data
         }
-        
+
         return gitClone(data.repoUrl, data.folderPath)
           .then(repo => {
             log({ type: 'info', message: 'Git repository cloned successuly.' })
