+++ lib/run.command.js
@@ -1,8 +1,8 @@
 const path = require('path')
 
-const { readFile, fileExist, chmod } = require('./fs.utils')
-const { execFile } = require('./process.utils')
-const { log } = require('./log.utils')
+const { readFile, fileExist, chmod } = require('./shared/fs.utils')
+const { execFile } = require('./shared/process.utils')
+const { log } = require('./shared/log.utils')
 
 module.exports = exports = configFilename => {
   return (scriptName, options) => {
@@ -21,7 +21,7 @@ module.exports = exports = configFilename => {
             return scriptPath
           }
         }
-        
+
         throw new Error('script_not_exist')
       })
       .then(scriptPath => chmod(scriptPath, '0700').then(() => scriptPath))
@@ -39,12 +39,12 @@ module.exports = exports = configFilename => {
         return child.toPromise()
       })
       .then(
-        () => log({ type: 'success', message: `"${scriptName}" script execution is finished.` }), 
+        () => log({ type: 'success', message: `"${scriptName}" script execution is finished.` }),
         error => {
           if (error.code === 'EACCES') {
             throw new Error('script_no_permission')
           }
-          
+
           throw error
         }
       )
