+++ lib/shared/fs.utils.js
