+++ lib/shared/git.utils.js
