+++ lib/shared/log.utils.js
